fix bugs  20231022 083508

+++ cfdns.js
@@ -2,9 +2,11 @@ const fs = require("fs");
 const dns = require('dns');
 const dnsPromises = dns.promises;
 const domains = fs.readFileSync('domains.txt','utf8').split('\n');
-fs.unlinkSync('./output/domainlist.txt');
-fs.unlinkSync('./output/domain2ip.txt');
-fs.unlinkSync('./output/domain2ip.txt.info');
+if (fs.existsSync('./output/domainlist.txt')) {
+    fs.unlinkSync('./output/domainlist.txt');
+    fs.unlinkSync('./output/domain2ip.txt');
+    fs.unlinkSync('./output/domain2ip.txt.info');
+  };
 (async function() {   
     for (const domain of domains){
         console.log(domain);
